requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.8.0,<3.0.0

